Suggestion for json structure
If using an object ("key": value, "key2": value2) instead of an array ([value, value2]) takes too much space in the end, we can drop all the keys for the events, provided we reconstruct it on the client's side once the data is downloaded. It's much nice to use objects than indexed arrays.